+++ perf/tests/test_perf_cli.py
@@ -233,8 +233,8 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             25.0 ms:  2 ###########
 
             Total duration: 16.0 sec
-            Start date: 2016-07-17T22:50:27
-            End date: 2016-07-17T22:50:46
+            Start date: 2016-07-17 22:50:27
+            End date: 2016-07-17 22:50:46
             Raw sample minimum: 96.9 ms
             Raw sample maximum: 100 ms
 
@@ -257,8 +257,8 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
     def test_stats(self):
         expected = ("""
             Total duration: 16.0 sec
-            Start date: 2016-07-17T22:50:27
-            End date: 2016-07-17T22:50:46
+            Start date: 2016-07-17 22:50:27
+            End date: 2016-07-17 22:50:46
             Raw sample minimum: 96.9 ms
             Raw sample maximum: 100 ms
 
@@ -325,7 +325,7 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             Run 1: warmup (1): 24.7 ms; samples (3): 24.5 ms, 24.5 ms, 24.5 ms
               cpu_freq: 1=3588 MHz
               cpu_temp: coretemp:Physical id 0=65 C, coretemp:Core 0=49 C, coretemp:Core 1=65 C
-              date: 2016-07-17T22:50:27
+              date: 2016-07-17 22:50:27
               load_avg_1min: 0.12
         """
         stdout = self.run_command('dump', '--verbose', TELCO)